Split the code which implements aligned buffer
