+++ packages/dev/core/src/Engines/engineFeatures.ts
@@ -69,9 +69,6 @@ export interface EngineFeatures {
     /**  Indicates that the engine supports sprite instancing */
     supportSpriteInstancing: boolean;
 
-    /** Indicates that the stride of a vertex buffer must always be a multiple of 4 bytes */
-    forceVertexBufferStrideMultiple4Bytes: boolean;
-
     /** @internal */
     _collectUbosUpdatedInFrame: boolean;
 }
