+++ packages/dev/core/src/Engines/nullEngine.ts
@@ -181,7 +181,6 @@ export class NullEngine extends Engine {
             needToAlwaysBindUniformBuffers: false,
             supportRenderPasses: true,
             supportSpriteInstancing: false,
-            forceVertexBufferStrideMultiple4Bytes: false,
             _collectUbosUpdatedInFrame: false,
         };
